@@ -11,10 +11,9 @@
     "@vueuse/components": "^10.9.0",
     "@vueuse/core": "^10.9.0",
     "bootstrap": "^4.1.1",
-    "bootstrap-vue": "^2.23.1",
-    "datatables.net-bs4": "^1.12.1",
-    "datatables.net-buttons-dt": "^2.2.3",
-    "datatables.net-dt": "^1.12.1",
+    "bootstrap-vue-next": "^0.16.6",
+    "datatables.net-bs4": "^2.0.3",
+    "datatables.net-buttons-bs4": "^3.0.1",
     "daterangepicker": "^2.1.25",
     "dygraphs": "^2.1.0",
     "filesize": "^10.1.0",
@@ -24,20 +23,20 @@
     "lodash": "^4.17.21",
     "moment": "^2.29.4",
     "popper.js": "^1.12.3",
-    "vue": "^2.7.10",
-    "vue-router": "^3.6.5",
+    "vue": "^3.4.21",
+    "vue-router": "^4.3.0",
     "vuedraggable": "^2.24.3"
   },
   "devDependencies": {
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
-    "@vitejs/plugin-vue2": "^1.1.2",
+    "@vitejs/plugin-vue": "^5.0.4",
     "autoprefixer": "^10.4.12",
     "husky": "^8.0.3",
     "lint-staged": "^14.0.1",
     "postcss": "^8.4.16",
     "prettier": "^3.0.3",
     "sass": "^1.54.5",
-    "vite": "^3.2.10"
+    "vite": "^5.2.7"
   },
   "lint-staged": {
     "*.{js,css,vue}": "prettier --write"
