@@ -1,4 +1,4 @@
-import vue from "@vitejs/plugin-vue2";
+import vue from "@vitejs/plugin-vue";
 import { defineConfig } from "vite";
 
 // https://vitejs.dev/config/
@@ -24,28 +24,22 @@ export default defineConfig({
     emptyOutDir: false,
     assetsDir: ".",
     rollupOptions: {
+      output: {
+        manualChunks: {
+          "highlight.js": ["highlight.js"],
+          datatables: ["datatables.net-bs4", "datatables.net-buttons-bs4"],
+          "bootstrap-vue-next": ["bootstrap-vue-next"],
+          vue: ["vue"],
+        },
+      },
       input: {
-        "alerting.checks": "/alerting.checks.js",
-        "alerting.check": "/alerting.check.js",
-        "settings.about": "/settings.about.js",
-        "settings.group": "/settings.group.js",
-        "settings.instance": "/settings.instance.js",
-        maintenance: "/maintenance.js",
-        "maintenance.database": "/maintenance.database.js",
-        "maintenance.schema": "/maintenance.schema.js",
-        "maintenance.table": "/maintenance.table.js",
-        activity: "/activity.js",
-        dashboard: "/dashboard.js",
         home: "/home.js",
-        monitoring: "/monitoring.js",
-        statements: "/statements.js",
-        temboard: "/temboard.js",
       },
     },
   },
   resolve: {
     alias: {
-      vue: "vue/dist/vue.esm.js",
+      vue: "vue/dist/vue.esm-bundler.js",
       moment: "moment/moment.js",
     },
   },
